Breaking up the search command into multiple commands based on type
